increase documents-per-block to 256
This allows use to process 256 documents in parallel (more or less).
After paying the cost to bring in a cache line, we use more of it before
discarding it.  BitFunnel processes an entire cache line at once (512
bits).  To make that effective in Go, we'll need assembly.

This improves query speed by 35%.