fix filter capacities for sharded index
We were rounding the length of the document terms to a power of two and
then size of the bloom filter.  We should instead just calculate the
correct power-of-two rounded bloom filter size of the given term list.

This fixed version uses less memory as smaller documents will be
assigned smaller bloom filters.